Fix s4 and s6 importer convertNewsQueue msvc build error

+++ src/openrct2/rct2/S6Importer.cpp
@@ -320,7 +320,7 @@ namespace OpenRCT2::RCT2
             return {};
         }
 
-        std::vector<OpenRCT2::News::Item> convertNewsQueue(const RCT12NewsItem* queue, uint8_t size)
+        std::vector<OpenRCT2::News::Item> convertNewsQueue(const RCT12NewsItem* queue, size_t size)
         {
             std::vector<OpenRCT2::News::Item> output{};
             const RCT12NewsItem* src = queue;
